requirements.txt
torch
torchaudio
demucs
openai-whisper
pydub
TTS
librosa
numpy
scipy
soundfile
